backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfminer.six==20231228
python-docx==1.1.2
pydantic==2.8.2
